clear_screenshots: check file extension

+++ Rakefile
@@ -6,8 +6,11 @@ desc "Clear Screenshot dir"
 task :clear_screenshots do
   screenshot_path = `defaults read com.apple.screencapture location`
   screenshot_path.delete!("\n")
+  
+  screenshot_type = `defaults read com.apple.screencapture type`
+  screenshot_type = ( screenshot_type.empty? )? ".png" : "." + screenshot_type.delete("\n")
 
-  if screenshot_path == "" || !File.directory?(screenshot_path)
+  if screenshot_path.empty? || !File.directory?(screenshot_path)
 
     log "remove_screenshots.log", Logger::ERROR, "No custom Screenshot DIR. I will not remove files from the desctop."
     exit -1
@@ -17,7 +20,9 @@ task :clear_screenshots do
   files_in_dir.each do |f|
     f_path = screenshot_path + "/" + f
     days_ago = (Date.today - 30).to_time
-    if File.ftype(f_path) == "file"
+    puts "Vor abfrage" + f_path
+    if File.ftype(f_path) == "file" && File.extname(f_path) == screenshot_type
+      puts "in Abfrage"
       if File.mtime(f_path) < days_ago
         begin
           File.delete f_path
